Restructure vscode-ws-jsonrpc sources: remove need for socket export
- Raise minimum engine level to node 22
