+++ mise.toml
@@ -1,5 +1,5 @@
 [tools]
-node = "24.16.0"
+node = "24"
 npm = "11"
 pnpm = "11"
 yarn = "4"
