+++ package.json
@@ -78,7 +78,7 @@
     "minimatch": "~10.2.5"
   },
   "engines": {
-    "node": ">=20.10.0",
-    "npm": ">=10.2.3"
+    "node": ">=22",
+    "npm": ">=10"
   }
 }
