+++ packages/client/package.json
@@ -124,7 +124,7 @@
     "monaco-languageclient-examples": "../examples"
   },
   "engines": {
-    "node": ">=20.10.0",
-    "npm": ">=10.2.3"
+    "node": ">=22",
+    "npm": ">=10"
   }
 }
