+++ packages/examples/resources/debugger/package.json
@@ -36,7 +36,7 @@
     "tsx": "~4.19.2"
   },
   "engines": {
-    "node": ">=20.10.0",
-    "npm": ">=10.2.3"
+    "node": ">=22",
+    "npm": ">=10"
   }
 }
