+++ packages/vscode-ws-jsonrpc/src/socket/connection.ts
@@ -3,9 +3,8 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import type { MessageConnection, Logger } from 'vscode-jsonrpc';
-import { createMessageConnection } from 'vscode-jsonrpc';
-import type { IWebSocket } from './socket.js';
+import { createMessageConnection, type Logger, type MessageConnection } from 'vscode-jsonrpc';
+import { ConsoleLogger, type IWebSocket } from 'vscode-ws-jsonrpc';
 import { WebSocketMessageReader } from './reader.js';
 import { WebSocketMessageWriter } from './writer.js';
 
@@ -15,4 +14,35 @@ export function createWebSocketConnection(socket: IWebSocket, logger: Logger): M
   const connection = createMessageConnection(messageReader, messageWriter, logger);
   connection.onClose(() => connection.dispose());
   return connection;
+}
+
+export function listen(options: { webSocket: WebSocket; logger?: Logger; onConnection: (connection: MessageConnection) => void }) {
+  const { webSocket, onConnection } = options;
+  const logger = options.logger ?? new ConsoleLogger();
+  webSocket.onopen = () => {
+    const socket = toSocket(webSocket);
+    const connection = createWebSocketConnection(socket, logger);
+    onConnection(connection);
+  };
+}
+
+export function toSocket(webSocket: WebSocket): IWebSocket {
+  return {
+    send: (content) => webSocket.send(content),
+    onMessage: (cb) => {
+      webSocket.onmessage = (event) => cb(event.data);
+    },
+    onError: (cb) => {
+      // oxlint-disable-next-line @typescript-eslint/no-explicit-any
+      webSocket.onerror = (event: any) => {
+        if (Object.hasOwn(event, 'message')) {
+          cb(event.message);
+        }
+      };
+    },
+    onClose: (cb) => {
+      webSocket.onclose = (event) => cb(event.code, event.reason);
+    },
+    dispose: () => webSocket.close()
+  };
 }
