--- packages/vscode-ws-jsonrpc/src/socket/index.ts
@@ -1,9 +1,0 @@
-/* --------------------------------------------------------------------------------------------
- * Copyright (c) 2024 TypeFox and others.
- * Licensed under the MIT License. See LICENSE in the package root for license information.
- * ------------------------------------------------------------------------------------------ */
-
-export * from './socket.js';
-export * from './reader.js';
-export * from './writer.js';
-export * from './connection.js';
