+++ packages/vscode-ws-jsonrpc/src/socket/reader.ts
@@ -3,10 +3,8 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-// TODO: Use environment-specific imports (vscode-jsonrpc/browser or vscode-jsonrpc/node)
-// when upgrading to vscode-jsonrpc@9.x.x-next.X which supports proper export maps
 import { AbstractMessageReader, type DataCallback, Disposable, MessageReader } from 'vscode-jsonrpc';
-import type { IWebSocket } from './socket.js';
+import type { IWebSocket } from 'vscode-ws-jsonrpc';
 
 export class WebSocketMessageReader extends AbstractMessageReader implements MessageReader {
   protected readonly socket: IWebSocket;
