+++ packages/vscode-ws-jsonrpc/tsconfig.src.json
@@ -1,6 +1,8 @@
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
+    // Workaround
+    "types": ["node"],
     "noEmit": false,
     "rootDir": "src",
     "outDir": "lib",
