+++ packages/wrapper-react/package.json
@@ -53,7 +53,7 @@
     "monaco-languageclient-examples": "../examples"
   },
   "engines": {
-    "node": ">=20.10.0",
-    "npm": ">=10.2.3"
+    "node": ">=22",
+    "npm": ">=10"
   }
 }
