+++ tsconfig.json
@@ -1,11 +1,9 @@
 {
   "compilerOptions": {
     "target": "ES2022",
-    "module": "node20",
+    "module": "nodenext",
     "moduleResolution": "nodenext",
     "lib": ["ES2022", "DOM", "DOM.Iterable"],
-    // "types": ["node"],
-    // esModuleInterop=true sets allowSyntheticDefaultImports=true,
     "esModuleInterop": true,
     "resolveJsonModule": true,
     // Enable project compilation
