+++ verify/next/package.json
@@ -31,7 +31,7 @@
     "@typefox/monaco-editor-react": "~8.0.0-next.0",
     "langium": "~4.3.0-next.a52b426",
     "monaco-languageclient": "~11.0.0-next.0",
-    "next": "~16.1.6",
+    "next": "~16.2.9",
     "react": "~19.2.7",
     "react-dom": "~19.2.7",
     "vscode": "npm:@codingame/monaco-vscode-extension-api@^33.0.9",
@@ -40,12 +40,12 @@
   },
   "devDependencies": {
     "@types/react-dom": "~19.2.3",
-    "@typescript/native-preview": "~7.0.0-dev.20260212.1",
+    "@typescript/native-preview": "beta",
     "shx": "~0.4.0",
     "vite": "~8.0.0-beta.14"
   },
   "engines": {
-    "node": ">=20.10.0",
-    "npm": ">=10.2.3"
+    "node": ">=22",
+    "npm": ">=10"
   }
 }
