+++ vite.config.ts
@@ -7,6 +7,7 @@ import { defineConfig } from 'vite';
 import fs from 'node:fs';
 import * as path from 'node:path';
 import vsixPlugin from '@codingame/monaco-vscode-rollup-vsix-plugin';
+import { fileURLToPath } from 'node:url';
 
 const clangdWasmLocation = 'packages/examples/resources/clangd/wasm/clangd.wasm';
 
@@ -54,16 +55,16 @@ export const definedViteConfig = defineConfig({
       '@codingame/monaco-vscode-standalone-html-language-features',
       '@codingame/monaco-vscode-standalone-json-language-features',
       '@codingame/monaco-vscode-standalone-typescript-language-features',
+      '@codingame/monaco-vscode-files-service-override',
       '@testing-library/react',
       'langium',
       'langium/lsp',
       'langium/grammar',
       'vscode/localExtensionHost',
       'vscode-jsonrpc',
-      'vscode-languageclient',
-      'vscode-languageserver',
-      'vscode-languageserver/browser',
-      'vscode-languageserver-protocol'
+      'vscode-languageclient/browser',
+      'vscode-languageserver-protocol/browser',
+      'vscode-languageserver/browser'
     ]
   },
   plugins: [
